+++ packages/jest-circus/src/legacy_code_todo_rewrite/jest_adapter_init.js
@@ -7,9 +7,9 @@
  * @flow
  */
 
-import type {TestResult, Status} from 'types/TestResult';
+import type {AssertionResult, TestResult, Status} from 'types/TestResult';
 import type {GlobalConfig, Path, ProjectConfig} from 'types/Config';
-import type {Event, TestEntry} from 'types/Circus';
+import type {Event, RunResult, TestEntry} from 'types/Circus';
 
 import {extractExpectedAssertionsErrors, getState, setState} from 'expect';
 import {formatExecError, formatResultsErrors} from 'jest-message-util';
@@ -19,12 +19,11 @@ import {
   buildSnapshotResolver,
 } from 'jest-snapshot';
 import {addEventHandler, dispatch, ROOT_DESCRIBE_BLOCK_NAME} from '../state';
-import {getTestID, getOriginalPromise} from '../utils';
+import {getTestID} from '../utils';
 import run from '../run';
 // eslint-disable-next-line import/default
 import globals from '../index';
 
-const Promise = getOriginalPromise();
 export const initialize = ({
   config,
   getPrettier,
@@ -123,46 +122,48 @@ export const runAndTransformResultsToJestFormat = async ({
   globalConfig: GlobalConfig,
   testPath: string,
 }): Promise<TestResult> => {
-  const runResult = await run();
+  const runResult: RunResult = await run();
 
   let numFailingTests = 0;
   let numPassingTests = 0;
   let numPendingTests = 0;
   let numTodoTests = 0;
 
-  const assertionResults = runResult.testResults.map(testResult => {
-    let status: Status;
-    if (testResult.status === 'skip') {
-      status = 'pending';
-      numPendingTests += 1;
-    } else if (testResult.status === 'todo') {
-      status = 'todo';
-      numTodoTests += 1;
-    } else if (testResult.errors.length) {
-      status = 'failed';
-      numFailingTests += 1;
-    } else {
-      status = 'passed';
-      numPassingTests += 1;
-    }
+  const assertionResults: Array<AssertionResult> = runResult.testResults.map(
+    testResult => {
+      let status: Status;
+      if (testResult.status === 'skip') {
+        status = 'pending';
+        numPendingTests += 1;
+      } else if (testResult.status === 'todo') {
+        status = 'todo';
+        numTodoTests += 1;
+      } else if (testResult.errors.length) {
+        status = 'failed';
+        numFailingTests += 1;
+      } else {
+        status = 'passed';
+        numPassingTests += 1;
+      }
 
-    const ancestorTitles = testResult.testPath.filter(
-      name => name !== ROOT_DESCRIBE_BLOCK_NAME,
-    );
-    const title = ancestorTitles.pop();
+      const ancestorTitles = testResult.testPath.filter(
+        name => name !== ROOT_DESCRIBE_BLOCK_NAME,
+      );
+      const title = ancestorTitles.pop();
 
-    return {
-      ancestorTitles,
-      duration: testResult.duration,
-      failureMessages: testResult.errors,
-      fullName: ancestorTitles.concat(title).join(' '),
-      invocations: testResult.invocations,
-      location: testResult.location,
-      numPassingAsserts: 0,
-      status,
-      title: testResult.testPath[testResult.testPath.length - 1],
-    };
-  });
+      return {
+        ancestorTitles,
+        duration: testResult.duration,
+        failureMessages: testResult.errors,
+        fullName: ancestorTitles.concat(title).join(' '),
+        invocations: testResult.invocations,
+        location: testResult.location,
+        numPassingAsserts: 0,
+        status,
+        title: testResult.testPath[testResult.testPath.length - 1],
+      };
+    },
+  );
 
   let failureMessage = formatResultsErrors(
     assertionResults,