+++ packages/jest-config/src/__tests__/normalize.test.js
@@ -11,7 +11,7 @@ import normalize from '../normalize';
 
 jest.mock('jest-resolve');
 
-jest.mock('path', () => require.requireActual('path').posix);
+jest.mock('path', () => jest.requireActual('path').posix);
 
 const crypto = require('crypto');
 const path = require('path');
@@ -330,7 +330,7 @@ describe('haste', () => {
   });
 });
 
-describe('setupTestFrameworkScriptFile', () => {
+describe('setupFilesAfterEnv', () => {
   let Resolver;
   beforeEach(() => {
     Resolver = require('jest-resolve');
@@ -344,36 +344,80 @@ describe('setupTestFrameworkScriptFile', () => {
     const {options} = normalize(
       {
         rootDir: '/root/path/foo',
-        setupTestFrameworkScriptFile: 'bar/baz',
+        setupFilesAfterEnv: ['bar/baz'],
       },
       {},
     );
 
-    expect(options.setupTestFrameworkScriptFile).toEqual(expectedPathFooBar);
+    expect(options.setupFilesAfterEnv).toEqual([expectedPathFooBar]);
   });
 
   it('does not change absolute paths', () => {
     const {options} = normalize(
       {
         rootDir: '/root/path/foo',
-        setupTestFrameworkScriptFile: '/an/abs/path',
+        setupFilesAfterEnv: ['/an/abs/path'],
       },
       {},
     );
 
-    expect(options.setupTestFrameworkScriptFile).toEqual(expectedPathAbs);
+    expect(options.setupFilesAfterEnv).toEqual([expectedPathAbs]);
   });
 
   it('substitutes <rootDir> tokens', () => {
     const {options} = normalize(
       {
         rootDir: '/root/path/foo',
-        setupTestFrameworkScriptFile: '<rootDir>/bar/baz',
+        setupFilesAfterEnv: ['<rootDir>/bar/baz'],
       },
       {},
     );
 
-    expect(options.setupTestFrameworkScriptFile).toEqual(expectedPathFooBar);
+    expect(options.setupFilesAfterEnv).toEqual([expectedPathFooBar]);
+  });
+});
+
+describe('setupTestFrameworkScriptFile', () => {
+  let Resolver;
+  let consoleWarn;
+
+  beforeEach(() => {
+    console.warn = jest.fn();
+    consoleWarn = console.warn;
+    Resolver = require('jest-resolve');
+    Resolver.findNodeModule = jest.fn(
+      name =>
+        name.startsWith('/') ? name : '/root/path/foo' + path.sep + name,
+    );
+  });
+
+  afterEach(() => {
+    console.warn = consoleWarn;
+  });
+
+  it('logs a deprecation warning when `setupTestFrameworkScriptFile` is used', () => {
+    normalize(
+      {
+        rootDir: '/root/path/foo',
+        setupTestFrameworkScriptFile: 'bar/baz',
+      },
+      {},
+    );
+
+    expect(consoleWarn.mock.calls[0][0]).toMatchSnapshot();
+  });
+
+  it('logs an error when `setupTestFrameworkScriptFile` and `setupFilesAfterEnv` are used', () => {
+    expect(() =>
+      normalize(
+        {
+          rootDir: '/root/path/foo',
+          setupFilesAfterEnv: ['bar/baz'],
+          setupTestFrameworkScriptFile: 'bar/baz',
+        },
+        {},
+      ),
+    ).toThrowErrorMatchingSnapshot();
   });
 });
 
@@ -804,6 +848,42 @@ describe('Upgrade help', () => {
   });
 });
 
+describe('testRegex', () => {
+  it('testRegex empty string is mapped to empty array', () => {
+    const {options} = normalize(
+      {
+        rootDir: '/root',
+        testRegex: '',
+      },
+      {},
+    );
+
+    expect(options.testRegex).toEqual([]);
+  });
+  it('testRegex string is mapped to an array', () => {
+    const {options} = normalize(
+      {
+        rootDir: '/root',
+        testRegex: '.*',
+      },
+      {},
+    );
+
+    expect(options.testRegex).toEqual(['.*']);
+  });
+  it('testRegex array is preserved', () => {
+    const {options} = normalize(
+      {
+        rootDir: '/root',
+        testRegex: ['.*', 'foo\\.bar'],
+      },
+      {},
+    );
+
+    expect(options.testRegex).toEqual(['.*', 'foo\\.bar']);
+  });
+});
+
 describe('testMatch', () => {
   it('testMatch default not applied if testRegex is set', () => {
     const {options} = normalize(
@@ -826,7 +906,7 @@ describe('testMatch', () => {
       {},
     );
 
-    expect(options.testRegex).toBe('');
+    expect(options.testRegex).toEqual([]);
   });
 
   it('throws if testRegex and testMatch are both specified', () => {
@@ -961,7 +1041,7 @@ describe('preset', () => {
 
   test('throws when preset is invalid', () => {
     jest.doMock('/node_modules/react-native/jest-preset.json', () =>
-      require.requireActual('./jest-preset.json'),
+      jest.requireActual('./jest-preset.json'),
     );
 
     expect(() => {
@@ -1215,7 +1295,7 @@ describe('testPathPattern', () => {
 
       describe('win32', () => {
         beforeEach(() => {
-          jest.mock('path', () => require.requireActual('path').win32);
+          jest.mock('path', () => jest.requireActual('path').win32);
           require('jest-resolve').findNodeModule = findNodeModule;
         });