+++ packages/jest-jasmine2/src/__tests__/p_timeout.test.js
@@ -16,14 +16,7 @@ describe('pTimeout', () => {
   it('calls `clearTimeout` and resolves when `promise` resolves.', async () => {
     const onTimeout = jest.fn();
     const promise = Promise.resolve();
-    await pTimeout(
-      promise,
-      Date.now(),
-      1000,
-      clearTimeout,
-      setTimeout,
-      onTimeout,
-    );
+    await pTimeout(promise, 1000, clearTimeout, setTimeout, onTimeout);
     expect(setTimeout).toHaveBeenCalled();
     expect(clearTimeout).toHaveBeenCalled();
     expect(onTimeout).not.toHaveBeenCalled();
@@ -33,14 +26,7 @@ describe('pTimeout', () => {
     const onTimeout = jest.fn();
     const promise = Promise.reject();
     try {
-      await pTimeout(
-        promise,
-        Date.now(),
-        1000,
-        clearTimeout,
-        setTimeout,
-        onTimeout,
-      );
+      await pTimeout(promise, 1000, clearTimeout, setTimeout, onTimeout);
     } catch (e) {}
     expect(setTimeout).toHaveBeenCalled();
     expect(clearTimeout).toHaveBeenCalled();
@@ -53,7 +39,6 @@ describe('pTimeout', () => {
     const promise = new Promise(() => {});
     const timeoutPromise = pTimeout(
       promise,
-      Date.now(),
       1000,
       clearTimeout,
       setTimeout,