+++ packages/jest-resolve/src/__tests__/is_builtin_module.test.js
@@ -1,3 +1,4 @@
+// Copyright (c) 2014-present, Facebook, Inc. All rights reserved.
 // @flow
 
 import isBuiltinModule from '../is_builtin_module';